requirements.txt
pyTelegramBotAPI
python-dotenv
SQLAlchemy
psycopg2-binary
APScheduler
pillow